Here's code that does much the same thing...
http://hintsforums.macworld.com/showpost.php?p=419594&postcount=16

I was also sorry to hear that cscreen had dropped off the Internet. In the hopes that future folks won't have to scrounge for bootleg copies, I've attached some sample code that does much the same as cscreen. This Macoshints page is the top Google hit on cscreen, so I thought it would be as good a place as any.

This code only sets resolution. Adding bit depth would be trivial for someone with interest. The command line parsing code is atrocious -- I'm not a programmer by day, so don't have the energy to spruce it up.

I compiled this code in XCode 3. I assume it will work in earlier versions. The APIs I used date back to OS 10.2, so if appropriately compiled it should be backward compatible at least that far. If you're going to compile this yourself, make sure to include the ApplicationServices framework in the project file.

Sample usage is:

newscreen -h 1440 -v 900

Hope this helps someone.